task_manager/requirements.txt
asgiref==3.8.1
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.2
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
pytest==8.3.4
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
django-environ==0.11.2
dj-database-url==2.3.0